requirements.txt
accelerate
appdirs
bitsandbytes
datasets
deepspeed
evaluate
fairscale
torch
gradio
huggingface-hub
loralib
nvitop
peft
sentencepiece
tensorboard
texttable
tokenizers
tqdm
transformers
flash-attn
wandb
triton
scipy
bert_score
jupyter
matplotlib
seaborn
nltk
spacy
jieba
rouge_score
tiktoken
transformers_stream_generator==0.0.4
einops
autoawq
vllm